queue: rename processor -> worker
This is just a small naming change to match the convention used.
